// production-export/src/models/interfaces.ts
export interface CreateLinkRequest {
  name: string;
  url: string;
  description?: string | undefined;
  iconUrl?: string | undefined;
  groupId: number;
  sortOrder?: number | undefined;
}